+++ bin/cli.js
@@ -5,7 +5,7 @@ var processTitle = "elm-verify-examples";
 process.title = processTitle;
 
 var argv = require('yargs').argv;
-var init = require('./modes').init;
+var init = require('./runner').init;
 
 
 // stateful things
